iommu/tegra: smmu: Fix u32 -> u64 for swgid
Fix u32 -> u64 for swgid.

Reported-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-on: http://git-master/r/213508

+++ drivers/iommu/tegra-smmu.c
@@ -370,7 +370,7 @@ static int __smmu_client_set_hwgrp(struct smmu_client *c, u64 map, int on)
 
 }
 
-static int smmu_client_set_hwgrp(struct smmu_client *c, u32 map, int on)
+static int smmu_client_set_hwgrp(struct smmu_client *c, u64 map, int on)
 {
 	u32 val;
 	unsigned long flags;
